Add a `scaffold` tool for easily debugging the scaffolder
The `dotnet-ef` and `ef` projects go through many hoops (`dotnet exec`) to run the scaffolder (`dotnet ef dbcontext scaffold`)
This makes it almost impossible to set a breakpoint and debug the code. This tools has the minimum code required to run the scaffolder and is easily debuggable with Visual Studio or Rider.
This tool was created so that I could investigate https://github.com/dotnet/efcore/issues/25729